+++ src/lib/data/constant.ts
@@ -12,7 +12,9 @@ import { MsgType } from "lib/types";
  */
 import { formatSlugName } from "lib/utils/format";
 
-export const DOCS_LINK = "https://docs.alleslabs.com/user-guide";
+export const USER_GUIDE_DOCS_LINK = "https://docs.alleslabs.com/user-guide";
+
+export const DEVELOPER_TOOL_DOCS_LINK = "https://docs.alleslabs.com/developers";
 
 export const INSTANTIATED_LIST_NAME = "Instantiated by me";