+++ src/lib/pages/proposal-details/components/vote-details/VoteDetailsAccordionItem.tsx
@@ -7,6 +7,7 @@ import {
 import type { ReactNode } from "react";
 
 import { ProposalStepper } from "../proposal-stepper";
+import { ProposalStatus } from "lib/types";
 import type { ProposalData } from "lib/types";
 
 interface VoteDetailsAccordionItemProps {
@@ -34,8 +35,14 @@ export const VoteDetailsAccordionItem = ({
     <AccordionPanel
       bg="transparent"
       p={0}
-      pt={3}
-      pb={step === 2 ? 4 : 0}
+      pt={
+        step === 1 &&
+        proposalData.status !== ProposalStatus.DEPOSIT_PERIOD &&
+        proposalData.status !== ProposalStatus.VOTING_PERIOD
+          ? 0
+          : 3
+      }
+      pb={step === 1 ? 0 : 4}
       borderTop="1px solid"
       borderColor="gray.700"
     >