+++ src/lib/pages/tx-details/components/TxInfo.tsx
@@ -76,8 +76,8 @@ export const TxInfo = ({ txData, ...flexProps }: TxInfoProps) => {
         )}
       </LabelText>
       <UserDocsLink
-        cta="More about Txs"
-        href="general/transaction/detail-page"
+        cta="Read more about Txs"
+        href="general/transactions/detail-page"
         mt={0}
       />
     </Container>